Updated .gitignore to exclude pycache
