+++ .gitignore
@@ -25,5 +25,14 @@ output/*
 src/__pycache__
 src/__pycache__/*
 
+src/core/__pycache__
+src/core/__pycache__/*
+
+src/gui/__pycache__
+src/gui/__pycache__/*
+
+src/utils/__pycache__
+src/utils/__pycache__/*
+
 tests/__pycache__
 tests/__pycache__/*
